pyproject.toml
[tool.poetry]
name = "topicaxis-classifiers"
version = "0.1.0"
description = "Topicaxis classifiers"
authors = ["Matigakis Panagiotis <pmatigakis@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
bentoml = "1.1.6"
scikit-learn = "1.2.2"
optuna = "3.1.1"
bcrypt = "3.2.2"
matplotlib = "3.6.1"
Pillow = "9.2.0"
pip = "^23.0.1"
llvmlite = "0.39.1"
numba = "0.56.4"
transformers = {version = "4.33.3", extras = ["torch"]}
datasets = "2.14.5"
evaluate = "0.4.0"
torchvision = "0.15.2"
torchaudio = "2.0.2"
nvidia-cudnn-cu11 = "8.9.4.25"
nvidia-cuda-cupti-cu11 = "11.7.101"
nvidia-nccl-cu11 = "2.18.3"
pip-tools = "7.3.0"
tensorflow = "2.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
